+++ src/pages/MeetingDetailPage.js
@@ -1,54 +1,11 @@
-import React, { Component } from 'react'
-import queryString from 'query-string'
+import React from 'react'
 
-import apis from 'apis'
+import MeetingDetail from 'component/Meeting/MeetingDetail'
 
-class MeetingDetailPage extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      meeting: null,
-    }
-  }
-
-  componentDidMount() {
-    const { match } = this.props
-    const meetingId = match.params.id
-    apis.readMeeting({meetingId}).then(value => this.setState({
-      meeting: value.data,
-    }))
-  }
-
-  toggleUserAttendanceHandler = user => {
-    const { meeting } = this.state
-    const meetingId = meeting.id
-    apis.toggleAttendance({ userId: user.id,  meetingId })
-        .then(() => apis.readMeeting({ meetingId }))
-        .then(value => this.setState({ meeting: value.data }))
-  }
-
-  isAttendance = user => {
-    const { meeting } = this.state
-    return meeting && meeting.attendances.includes(user.id)
-  }
-
-  render() {
-    const { meeting } = this.state
-    return (meeting &&
-        <div>
-          Meeting Time: {meeting.time}<br/>
-          Attendance
-          {meeting.group.members.map((user, index) =>
-              <div key={index}>
-                {user.username}
-                <button onClick={() => this.toggleUserAttendanceHandler(user)}>
-                  {this.isAttendance(user) ? 'O' : 'X'}
-                </button>
-              </div>
-          )}
-        </div>
-    )
-  }
-}
+const MeetingDetailPage = props => (
+  <div>
+    <MeetingDetail match = {props.match}/>
+  </div>
+)
 
 export default MeetingDetailPage
