+++ src/routes.js
@@ -1,5 +1,6 @@
 export default {
   LOGIN: '/login',
+  REGISTER: '/register',
   GROUP_LIST: '/groups',
   GROUP_FORM: '/groups/new',
   GROUP_DETAIL: '/groups/:id',
